requirements.txt
streamlit==1.40.0
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
scipy==1.14.1
implicit==0.7.2
scikit-surprise==1.1.4
Pillow==11.0.0
requests==2.32.3
tqdm==4.66.6
